requirements.txt
torch>=2.3
torchvision>=0.18
numpy>=1.24
pillow>=10.0
opencv-python>=4.9
scipy>=1.11
scikit-image>=0.22
safetensors>=0.4
pyyaml>=6.0
tqdm>=4.65
tensorboard>=2.16
lpips>=0.1.4
imagehash>=4.3
fastapi>=0.110
uvicorn>=0.27
pytest>=8.0
